r55: Improve security; change group instead of user on HTTP.keytab
There was a possibility that the log file used by setup-mod_auth_vas could
be used in a symlink-race attack on /tmp. The script now uses noclobber.

There was a possibility that a compromised web server may be able to 
chmod and inject data into the HTTP.keytab file. The keytab file is now
left as owned by its creator (usually root), and read access granted to 
the web server via its group membership.