r199: Add -u flag, shell portability, intelligently default SPN additions.
The -u flag allows the script to pass -u to vastool when needed.
Otherwise "-u HTTP/" is used where possible, thus allowing for re-use of
the keytab for low privilege queries. If -u is not provided, then the existing
krb5cc is used, or kinit is called.

Binary AND and OR functions are supplied for where $((..)) syntax is
unavailable.

When prompting for SPNs to add, use the samAccountName$ of host/ to determine
the alias "HTTP/<samAccountName>" and set that as the initial default when
adding aliases.