r301: Use RETURN() instead of user_matches variables
Mainly because it fixes a case where the return value of match_container
might return some non-match value when it should have been a match
depending on uninitialized memory.

Fortunately this time it's only buggy in the user-is-authorised case.
But I'm annoyed that GCC only noticed this when compiling with APXS1. WTF?

Fixes bug #450.