Don't set flag variables to empty It's not the same as not being set at all
eg. if uflag is set but empty, this evaluates true and substitutes
the text on the right:
  ${uflag+-u "$uflag"}
which results in just "-u " and causes Bad Things to happen.
On the other hand if uflag is not set (eg by `unset`) it does what you'd
hope, except that in bug #504 we found out that HP-UX's shell chokes on
`unset`.

Fixes bug #525.