setup-mod_auth_vas4.in: also support rerunning the script if it fails
If the HTTP service keytab was already created successfully, but some
other parts fail (eg checking that the module is loaded), it might be
a reasonable thing to rerun the script (after fixing the problem manually)
and revalidate.

However, if the keytab exists, we do not fill some variables which are
required for the sanity checks to succeed (HTTP_SPN, SAMACCOUNTNAME),
which makes the calls fail. Previously the only workaround was to remove
the keytab, now the script asks for the required info and is able to
finish (if the keytab was correct).