setup-mod_auth_vas4.in: optionally let user enable vas module loading
On ubuntu/debian at least, the apache configuration has a directory scheme
like:

/etc/apache2
   \--- mods-available
          \--- auth_vas4.load
   \--- mods-enabled

We install the module under the available modules, but it is not enabled
by default. For that, we need to symlink it under the mods-enabled
directory. So if it is not done, checking the module loading will fail.

While this can be done by the user manually before calling the setup
script, I think it is fun if we ask him and
do so automatically (but of course only if the directory scheme gets
detected). Since this config structure can be elsewhere and can have
other format, we just skip this if we do not find this structure.