configure.ac: skip unsupported cflags from apxs
If apache was compiled with a different compiler, apxs may report cflags
which are not supported by the compiler we use. Since converting these
between compilers seems impossible, our strategy here would be to skip
these flags and warn the user. (S)He can still add the missing CFLAGS
to the configure call if required.