log.h: sync variadic logging macros with their GNUC alternatives
to make the code compile with non gnuc compiler. (In my case Oracle
Solaris Studio.)

This code path was not compiling, I guess because we mainly use gnuc
compatible compilers. The difference with __VA_ARGS__ and ##a scheme
is just that __VA_ARGS__ does not support being empty. So it needs to
replace one more argument to be successful.